refactor(planning): move split modules into planning.engine/ (directory convention)
Reorganization by @fabrice8: the planning split now mirrors the
executive.engine/ directory convention — planning.engine/{engine,types,
plan.store,plan.frontier,plan.supervision}.ts replaces the flat
planning.engine.ts + planning/ pair. Importers updated across src + tests.

Gates: typecheck ✓ · suite 892/1/0 ✓ · replay capstone ✓ · dist rebuilt.
