+++ src/cognition/faculties/executive.engine/engine.ts
@@ -32,7 +32,7 @@ import type { WorkingMemory } from '#faculties/working.memory'
 import type { GoalManager, GoalState } from '#faculties/goal.manager'
 import type { EpisodicConsolidator } from '#faculties/episodic.consolidator'
 import type { SemanticIntegrator, Belief } from '#faculties/semantic.engine'
-import type { PlanningEngine } from '#faculties/planning.engine'
+import type { PlanningEngine } from '#cognition/faculties/planning.engine/engine'
 import type { TokenTracker } from '#cognition/utilities/token.tracker'
 import type { CognitiveEngine } from '#cognition/types'
 import type { CognitiveEvent, CognitiveBus } from '#cognition/bus'
