+++ tests/integration/mind.integration.test.ts
@@ -43,7 +43,7 @@ import { DreamSimulator } from '#faculties/dream.simulator'
 import { GoalManager } from '#faculties/goal.manager'
 import { InhibitionController } from '#faculties/inhibition.controller'
 import { TaskSwitcher } from '#faculties/task.switcher'
-import { PlanningEngine } from '#faculties/planning.engine'
+import { PlanningEngine } from '#cognition/faculties/planning.engine/engine'
 import { SelfModelUpdater } from '#faculties/self.model.updater'
 import { ConfidenceCalibrator } from '#faculties/confidence.calibrator'
 import { BiasDetector } from '#faculties/bias.detector'
