+++ tests/unit/planning.disposition.test.ts
@@ -10,7 +10,7 @@
  */
 
 import { describe, it, expect } from 'vitest'
-import { PlanningEngine } from '#faculties/planning.engine'
+import { PlanningEngine } from '#cognition/faculties/planning.engine/engine'
 import { createTestBus }  from '#cognition/bus'
 
 const step = () => ( { action: 'observe', description: 'd', expectedOutcome: 'o', prerequisites: [], estimatedDuration: 3 } )
