+++ tests/unit/planning.emergent-tier.test.ts
@@ -9,7 +9,7 @@
  */
 
 import { describe, it, expect } from 'vitest'
-import { PlanningEngine } from '#faculties/planning.engine'
+import { PlanningEngine } from '#cognition/faculties/planning.engine/engine'
 import { createTestBus }  from '#cognition/bus'
 
 const makeState = ( tick: number ) => ( { tick, metrics: new Map(), entities: new Map() } as any )
