+++ tests/unit/planning.execution.test.ts
@@ -15,7 +15,7 @@
  */
 
 import { describe, it, expect, beforeEach } from 'vitest'
-import { PlanningEngine }  from '#faculties/planning.engine'
+import { PlanningEngine }  from '#cognition/faculties/planning.engine/engine'
 import { createTestBus }   from '#cognition/bus'
 
 // ── Minimal collaborators ─────────────────────────────────────
