+++ tests/unit/planning.goal-sync.test.ts
@@ -10,7 +10,7 @@
  */
 
 import { describe, it, expect } from 'vitest'
-import { PlanningEngine } from '#faculties/planning.engine'
+import { PlanningEngine } from '#cognition/faculties/planning.engine/engine'
 import { GoalManager }    from '#faculties/goal.manager'
 import { createTestBus }  from '#cognition/bus'
