+++ tests/unit/planning.multiplan.test.ts
@@ -11,7 +11,7 @@
  */
 
 import { describe, it, expect, beforeEach } from 'vitest'
-import { PlanningEngine }      from '#faculties/planning.engine'
+import { PlanningEngine }      from '#cognition/faculties/planning.engine/engine'
 import { createTestBus }       from '#cognition/bus'
 import { buildExecutiveContext } from '#faculties/executive.engine/context'
