+++ tests/unit/planning.vocabulary.test.ts
@@ -8,7 +8,7 @@
  */
 
 import { describe, it, expect } from 'vitest'
-import { PlanningEngine } from '#faculties/planning.engine'
+import { PlanningEngine } from '#cognition/faculties/planning.engine/engine'
 import { createTestBus }  from '#cognition/bus'
 
 const makeState = ( t: number ) => ( { tick: t, metrics: new Map(), entities: new Map() } as any )
